feat: Add getHappyHour API endpoint and related functions
This commit adds a new API endpoint and related functions to retrieve happy hour details based on the provided ID. The `getHappyHour` function in the `api-path.ts` file is implemented to return the encoded URI for the happy hour endpoint. Additionally, the `toStringHour` and `extractHourMinutesDate` functions are added to the `date.format.ts` file to handle hour formatting and extraction.